+++ router/internal/controlplane/configpoller/config_poller.go
@@ -14,7 +14,6 @@ import (
 	"github.com/wundergraph/cosmo/router/gen/proto/wg/cosmo/node/v1/nodev1connect"
 	"github.com/wundergraph/cosmo/router/internal/cdn"
 	"github.com/wundergraph/cosmo/router/internal/controlplane"
-	"github.com/wundergraph/cosmo/router/internal/jwt"
 	"go.uber.org/zap"
 	brotli "go.withmatt.com/connect-brotli"
 )
@@ -126,13 +125,7 @@ func (c *configPoller) Subscribe(ctx context.Context, handler func(newConfig *no
 func (c *configPoller) getRouterConfigFromCP(ctx context.Context, version *string) (*nodev1.RouterConfig, error) {
 	start := time.Now()
 
-	claims, err := jwt.ExtractFederatedGraphTokenClaims(c.graphApiToken)
-	if err != nil {
-		return nil, err
-	}
-
 	req := connect.NewRequest(&nodev1.GetConfigRequest{
-		GraphId: claims.FederatedGraphID,
 		Version: version,
 	})