+++ shared/CHANGELOG.md
@@ -4,6 +4,16 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.19.0](https://github.com/wundergraph/cosmo/compare/@wundergraph/cosmo-shared@0.18.4...@wundergraph/cosmo-shared@0.19.0) (2024-01-30)
+
+### Features
+
+* implement authorization directives ([#448](https://github.com/wundergraph/cosmo/issues/448)) ([181d89d](https://github.com/wundergraph/cosmo/commit/181d89d8e7dbf8eb23cddfa0b6c91c840a2986b0)) (@Aenimus)
+
+## [0.18.4](https://github.com/wundergraph/cosmo/compare/@wundergraph/cosmo-shared@0.18.3...@wundergraph/cosmo-shared@0.18.4) (2024-01-29)
+
+**Note:** Version bump only for package @wundergraph/cosmo-shared
+
 ## [0.18.3](https://github.com/wundergraph/cosmo/compare/@wundergraph/cosmo-shared@0.18.2...@wundergraph/cosmo-shared@0.18.3) (2024-01-26)
 
 **Note:** Version bump only for package @wundergraph/cosmo-shared